Rename modules for persistance across the project.
