--- backend/services/.gitignore
@@ -1,1 +1,0 @@
-target
