--- backend/services/Ballerina.lock
@@ -1,9 +1,0 @@
-org_name = "yashodperera"
-version = "0.1.0"
-lockfile_version = "1.0.0"
-ballerina_version = "1.1.0"
-
-[[imports."yashodperera/admin_services:0.1.0"]]
-org_name = "yashodperera"
-name = "utilities"
-version = "0.1.0"
