--- backend/services/Ballerina.toml
@@ -1,5 +1,0 @@
-[project]
-org-name= "yashodperera"
-version= "0.1.0"
-
-[dependencies]
