--- backend/services/src/admin_services/tests/main_test.bal
@@ -1,39 +1,0 @@
-import ballerina/io;
-import ballerina/test;
-
-# Before Suite Function
-
-@test:BeforeSuite
-function beforeSuiteFunc() {
-    io:println("I'm the before suite function!");
-}
-
-# Before test function
-
-function beforeFunc() {
-    io:println("I'm the before function!");
-}
-
-# Test function
-
-@test:Config {
-    before: "beforeFunc",
-    after: "afterFunc"
-}
-function testFunction() {
-    io:println("I'm in test function!");
-    test:assertTrue(true, msg = "Failed!");
-}
-
-# After test function
-
-function afterFunc() {
-    io:println("I'm the after function!");
-}
-
-# After Suite Function
-
-@test:AfterSuite
-function afterSuiteFunc() {
-    io:println("I'm the after suite function!");
-}
