--- backend/services/src/user_services/constants.bal
@@ -1,9 +1,0 @@
-// Client APIs' section.
-const string GITHUB_API_URL = "https://api.github.com";
-const string ORGANIZATION_NAME = "yashodgayashan";
-const string REPOSITORY_NAME = "ballerina-github-connector";
-const string ACCESS_TOKEN = "Bearer <token>";
-
-// Listner configurations.
-const int USER_SERVICES_PORT = 9080;
-const string BASEPATH = "/user-services";
