--- backend/services/src/utilities/Module.md
@@ -1,5 +1,0 @@
-Prints "Hello World!" with a main function.
-[//]: # (above is the module summary)
-
-# Module Overview
-Prints "Hello World!" as the output to the command line using a main function.
