+++ frontend/web/src/components/home/body/dashboard/admin_dash/request_info/request_info.js
@@ -1,5 +1,5 @@
 import React, { Fragment, useContext, useEffect, useState } from "react";
-import { AdminRequestContext } from "../../../../../utility/contexts/adminRequestContext";
+import { AdminRequestContext } from "../../../../../utility/contexts/admin_request_context";
 import "./request_info.css";
 import { useHistory } from "react-router-dom";
 import axios from "axios";
