+++ frontend/web/src/components/home/body/dashboard/user_dash/request_info/request_info.js
@@ -1,6 +1,6 @@
 import React, { Fragment, useContext, useEffect, useState } from "react";
 import { useHistory } from "react-router-dom";
-import { RequestContext } from "../../../../../utility/contexts/requestContext";
+import { RequestContext } from "../../../../../utility/contexts/request_context";
 import "./request_info.css";
 import axios from "axios";
 import { Button, Comment, Form, Header } from "semantic-ui-react";
