+++ frontend/web/src/components/home/body/login/login.css
@@ -1,5 +1,5 @@
 .background {
-  background-image: url("./background.png");
+  background-image: url("../../../../asserts/images/background.png");
   background-repeat: no-repeat;
   background-size: cover;
   height: 830px;
