--- frontend/web/src/components/home/header/header.css
@@ -1,26 +1,0 @@
-.header {
-  position: fixed;
-  left: 0;
-  top: 0;
-  padding: 10px;
-  width: 100%;
-  height: 5vh;
-  z-index: 3;
-  background-color: rgb(215, 226, 211);
-  color: rgb(76, 175, 80);
-}
-
-.header-logo {
-  display: inline;
-  color: rgb(253, 122, 0);
-}
-
-.header-button {
-  float: right;
-}
-
-.header-header {
-  position: fixed;
-  left: 3%;
-  top: 10px;
-}
