+++ frontend/web/src/components/home/home.js
@@ -12,9 +12,9 @@ import LoginForm from "../home/body/login/login";
 import SignupForm from "../home/body/signup/signup";
 import User from "./body/dashboard/user_dash/user_dash";
 import Admin from "./body/dashboard/admin_dash/admin_dash";
-import Notification from "../Notification/notification";
+import Notification from "../notification/notification";
 import Request from "./body/dashboard/user_dash/request_info/request_info";
-import LandRequestForm from "../LandRequestForm/land-request-form";
+import LandRequestForm from "./body/request_form/land-request-form";
 import AdminRequest from "./body/dashboard/admin_dash/request_info/request_info";
 import "./home.css";
 
@@ -30,9 +30,9 @@ const Home = () => {
             <Route exact path="/signup" component={SignupForm} />
             <Route exact path="/user-dash" component={User} />
             <Route exact path="/admin-dash" component={Admin} />
-            <Route exact path="/notifications" component={Notification} />
             <Route exact path="/user-dash/request" component={Request} />
             <Route exact path="/admin-dash/request" component={AdminRequest} />
+            <Route exact path="/notifications" component={Notification} />
             <Route exact path="/new-issue" component={LandRequestForm} />
             <Redirect exact from="" to="/login" />
           </Switch>
